+++ src/meta_usage.rs
@@ -118,7 +118,7 @@ fn collect_usage_meta(meta: &Meta, is_pos: &mut bool) -> Option<UsageMeta> {
             }
         }
         Meta::Decorated(meta, _) => collect_usage_meta(meta, is_pos)?,
-        Meta::Skip => return None,
+        Meta::HideUsage(_) | Meta::Skip => return None,
         Meta::Item(i) => match i {
             Item::Positional {
                 metavar, strict, ..