+++ tests/meta_youmean.rs
@@ -2,6 +2,7 @@ use bpaf::*;
 
 #[test]
 fn ambiguity() {
+    set_override(false);
     #[derive(Debug, Clone)]
     enum A {
         V(Vec<bool>),
@@ -27,6 +28,7 @@ fn ambiguity() {
 
 #[test]
 fn short_cmd() {
+    set_override(false);
     let parser = long("alpha")
         .req_flag(())
         .to_options()